+++ src/appmake/kc.c
@@ -62,7 +62,7 @@ int kc_exec(char* target)
     // strupr(filename);
     // not available on all platforms
 
-    for (i = strlen(filename) - 1; i >= 0 && filename[i] != '/' && filename[i] != '\\'; i--)
+    for (i = (int)strlen(filename) - 1; i >= 0 && filename[i] != '/' && filename[i] != '\\'; i--)
         filename[i] = toupper(filename[i]);
 
     //
